requirements.txt
pytesseract>=0.3.10
opencv-python>=4.8.0
Pillow>=10.0.0
watchdog>=3.0.0
requests>=2.31.0
numpy>=1.24.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
prometheus-client>=0.19.0
boto3>=1.34.0
pydantic>=2.5.0